create responsive app shell wrapper
